Move messages prints to a print file and create exit commands
