+++ golings/cmd/print.go
@@ -1,0 +1,72 @@
+package cmd
+
+import (
+	"os"
+	"os/exec"
+	"runtime"
+
+	"github.com/fatih/color"
+	"github.com/mauricioabreu/golings/golings/exercises"
+	"github.com/mauricioabreu/golings/golings/ui"
+)
+
+func PrintHint(infoFile string) {
+	ClearScreen()
+	exercise, err := exercises.NextPending(infoFile)
+	if err != nil {
+		color.Red("Failed to find next exercises")
+	}
+	color.Yellow(exercise.Hint)
+}
+
+func PrintList(infoFile string) {
+	ClearScreen()
+	exs, err := exercises.List(infoFile)
+	if err != nil {
+		color.Red("Failed to list exercises")
+	}
+	ui.PrintList(os.Stdout, exs)
+}
+
+func RunNextExercise(infoFile string) {
+	ClearScreen()
+	exercise, err := exercises.NextPending(infoFile)
+	if err != nil {
+		color.Red("Failed to find next exercises")
+	}
+
+	result, err := exercise.Run()
+	if err != nil {
+		color.Cyan("Failed to compile the exercise %s\n\n", result.Exercise.Path)
+		color.White("Check the output below: \n\n")
+		color.Red(result.Err)
+		color.Red(result.Out)
+		color.Yellow("If you feel stuck, ask a hint by executing `golings hint %s`", result.Exercise.Name)
+	} else {
+		color.Green("Congratulations!\n\n")
+		color.Green("Here is the output of your program:\n\n")
+		color.Cyan(result.Out)
+		if result.Exercise.State() == exercises.Pending {
+			color.White("Remove the 'I AM NOT DONE' from the file to keep going\n")
+			color.Red("exercise is still pending")
+		}
+	}
+}
+
+func ClearScreen() {
+	if runtime.GOOS == "windows" {
+		cmd := exec.Command("cmd", "/c", "cls")
+		cmd.Stdout = os.Stdout
+		err := cmd.Run()
+		if err != nil {
+			color.Red("Clear terminal command error")
+		}
+	} else {
+		cmd := exec.Command("clear")
+		cmd.Stdout = os.Stdout
+		err := cmd.Run()
+		if err != nil {
+			color.Red("Clear terminal command error")
+		}
+	}
+}
